django_E-commerce-main/requirements.txt
Django>=4.2.7
Pillow>=10.0.0
python-dotenv>=1.0.0
sqlparse>=0.4.2
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
Werkzeug==2.0.2
WTForms==3.0.0
whitenoise==5.3.0
gunicorn==20.1.0
social-auth-app-django>=5.0.0
social-auth-core>=4.1.0
python3-openid>=3.2.0
requests-oauthlib>=1.3.0
oauthlib>=3.1.1
mysqlclient>=2.2.0